Add locks to scraper rate limiter
The file requests seem to be starving the image requests somewhat; which
doesn't really matter in the end, but might as well.

Still kind of weird that the sleeps aren't fair either, but oh well. I
roughly guess it's because when requests complete, they can mess with
the predicted rate, so the sleeps don't all line up without this lock.